+++ packages/ui/src/Toast/index.ts
@@ -1,0 +1,2 @@
+export * from "./store";
+export { default as Toast } from "./Toast";
