+++ packages/utils/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./hooks";
-export * from "./store";
 export * from "./utils";
